Backend/requirements.txt
# fastapi
# uvicorn
# python-multipart
pillow
torch
numpy
kaggle
transformers
datasets
scikit-learn
evaluate
mlflow
pandas
accelerate
optuna
torchvision
imbalanced-learn
